+++ server/src/middleware/auth.js
@@ -1,6 +1,9 @@
+import bodyParser from 'body-parser';
 import session from 'express-session';
 
-const sessionMiddleware = session({
+export const bodyParse = bodyParser.urlencoded({ extended: true });
+
+export const sessionCookie = session({
     secret: process.env.SESSION_SECRET,
     resave: false,
     saveUninitialized: true,
@@ -10,4 +13,3 @@ const sessionMiddleware = session({
         secure: false,
     },
 });
-export default sessionMiddleware;
