--- server/src/middleware/bodyparser.js
@@ -1,5 +1,0 @@
-import bodyParser from 'body-parser';
-
-const bodyParserMiddleware = bodyParser.urlencoded({ extended: true });
-export default bodyParserMiddleware;
-
