+++ server/src/middleware/cors.js
@@ -1,8 +1,7 @@
 import cors from 'cors';
 
-const corsMiddleware = cors({
+export const corsMiddleware = cors({
     origin: process.env.FRONTEND_URL,
     credentials: true,
 });
-export default corsMiddleware;
